backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
pyjwt
python-jose[cryptography]